Paul - 10/20/2012 Various changes and bug-fixes - Timings for RandomAssessment and Drinking follow ups have been reduced for testing, don't forget to change them back for actual use.
Survey XML Files
- Fixed various typos and formatting problems.

Survey
- Increased text size of questions.
- Fixed drinking follow-up incorrectly triggering.  Pending intents did
not have unique request codes so they were overriding each other.
- Removed unnecessary logging in Handler and ConfigHandler.
- Removed extra call to SurveyQuestion.getSelectedAnswers() for the last
question in a survey in XMLSurveyActivity.

Service
- Fixed DrinkingFollowup survey being triggered twice.
- Fixed null pointer exception in LocationControl, location recording
seems to be working but there may be additional bugs.

Known Issues:
- SensorControl isn't working, data appears in database as NaN, program
doesn't crash or anything so an exception is being ignored or the
control isn't recording data correctly.  Could be in either
SensorService, SensorControl or super-class.